requirements.txt
pymongo==4.1.1
azure-eventhub==5.10.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
redis==4.3.3
azure-storage-blob==12.12.0
azure-servicebus==7.7.0
